+++ code-types/TypeConverter.js
@@ -7,26 +7,28 @@ var TypeConverter;
     /** Convert a CodeAst.Type to a string in the format 'TypeName<GenericParams, ...>[][]...'
      * @param type the type to stringify
      * @param [typeConverter] an optional converter for the type names in the type
+     * @param [includeNullability=true] whether to include type nullability (i.e. the '?' part of 'int?[]')
      */
-    function typeToString(type, typeConverter) {
+    function typeToString(type, typeConverter, includeNullability) {
+        if (includeNullability === void 0) { includeNullability = true; }
         var dst = [];
-        _typeToString(type, dst, typeConverter);
+        _typeToString(type, dst, typeConverter, includeNullability);
         return dst.join("");
     }
     TypeConverter.typeToString = typeToString;
-    function _typeToString(type, dst, typeConverter) {
+    function _typeToString(type, dst, typeConverter, includeNullability) {
         dst.push(typeConverter ? typeConverter(type.typeName) : type.typeName);
         var childs = type.genericParameters;
         if (childs && childs.length > 0) {
             dst.push("<");
             for (var i = 0, sizeM1 = childs.length - 1; i < sizeM1; i++) {
-                _typeToString(childs[i], dst, typeConverter);
+                _typeToString(childs[i], dst, typeConverter, includeNullability);
                 dst.push(", ");
             }
-            _typeToString(childs[sizeM1], dst, typeConverter);
+            _typeToString(childs[sizeM1], dst, typeConverter, includeNullability);
             dst.push(">");
         }
-        if (type.nullable) {
+        if (type.nullable && includeNullability) {
             dst.push("?");
         }
         if (type.arrayDimensions) {
@@ -179,22 +181,25 @@ var TypeConverter;
          * If given a Type, convert the types to TypeScript and convert to a string (see typeToString()).
          * @param typeTemplate a Type, or string where the format must be 'typeName?[][]...' where typeName has no generic parameters, and the '?' (nullability) and '[][]...' (array dimensions) are optional
          * @param returnUnknownTypes
+         * @param [includeNullability=false]
          */
-        TypeScript.parseAndConvertTypeTemplate = function (typeTemplate, returnUnknownTypes) {
+        TypeScript.parseAndConvertTypeTemplate = function (typeTemplate, returnUnknownTypes, includeNullability) {
+            if (includeNullability === void 0) { includeNullability = false; }
             return (typeof typeTemplate === "string"
-                ? TypeScript.parseAndConvertTypeTemplateString(typeTemplate, returnUnknownTypes)
-                : typeToString(typeTemplate, function (t) { return TypeScript.convertSimpleType(t, returnUnknownTypes); }));
+                ? TypeScript.parseAndConvertTypeTemplateString(typeTemplate, returnUnknownTypes, includeNullability)
+                : typeToString(typeTemplate, function (t) { return TypeScript.convertSimpleType(t, returnUnknownTypes); }, includeNullability));
         };
         /** Parse and convert a simple type template string to a TypeScript type string.
          * The format must be 'typeName?[][]...' where typeName has no generic parameters, and the '?' (nullability) and '[][]...' (array dimensions) are optional
          * @param typeName the format must be 'typeName?[][]...' where typeName has no generic parameters, and the '?' (nullability) and '[][]...' (array dimensions) are optional
          * @param returnUnknownTypes
+         * @param [includeNullability=false]
          */
-        TypeScript.parseAndConvertTypeTemplateString = function (typeTemplate, returnUnknownTypes) {
+        TypeScript.parseAndConvertTypeTemplateString = function (typeTemplate, returnUnknownTypes, includeNullability) {
             var typeInfo = TypeConverter.parseTypeTemplate(typeTemplate);
             var arrayCount = typeInfo.arrayDimensions;
             var tsType = TypeScript.convertSimpleType(typeInfo.typeName, returnUnknownTypes);
-            return tsType + (typeInfo.nullable ? "?" : "") + (arrayCount > 0 ? new Array(arrayCount + 1).join("[]") : "");
+            return tsType + (typeInfo.nullable && includeNullability ? "?" : "") + (arrayCount > 0 ? new Array(arrayCount + 1).join("[]") : "");
         };
         /** Convert primitive and common builtin C# and Java types to TypeScript equivalent types
          * @param typeName the type name (i.e. 'bool' or 'String')